requirements.txt
gradio==4.38.1
assemblyai==0.30.0
translate==3.6.1
elevenlabs==1.4.1
deepgram-sdk==3.4.0
deep-translator==1.11.4
python-dotenv